requirements.txt
# --- LlamaIndex 核心与 Gemini 组件 ---
llama-index-core
llama-index-llms-google-genai
llama-index-embeddings-google-genai
google-genai

# --- 向量数据库 ---
llama-index-vector-stores-qdrant
qdrant-client

# --- 文档解析 (Docling) ---
llama-index-readers-docling
docling

# --- 重排序 (Rerank) ---
llama-index-postprocessor-sbert-rerank
sentence-transformers

# --- PDF 处理与截图 ---
pymupdf
pillow

# --- Web 界面与工具 ---
streamlit
pandas
python-dotenv
nest_asyncio